+++ src/main/java/com/example/fastest_server/variant/VariantRepository.java
@@ -11,16 +11,16 @@ import java.util.List;
 
 @Repository
 @Transactional
-public interface VariantRepository extends JpaRepository<Variant, Integer> {
+public interface VariantRepository extends JpaRepository<Variant, Long> {
 
     @Query(value = "SELECT * FROM variants WHERE id IN (" +
             "SELECT variant_id FROM variant_question WHERE question_id IN " +
             "(SELECT id FROM questions WHERE test_id = :idTest)) ORDER BY student_name", nativeQuery = true)
 
-    List<Variant> findByTestId(@Param("idTest") int idTest);
+    List<Variant> findByTestId(@Param("idTest") Long idTest);
 
     @Modifying
     @Query(value = "update variants set mark = :mark where id= :idVariant", nativeQuery = true)
-    void setMarkToVariant (@Param("idVariant") int idVariant, @Param("mark") int mark);
+    void setMarkToVariant (@Param("idVariant") Long idVariant, @Param("mark") Integer mark);
 
 }
