+++ src/main/java/com/example/fastest_server/variant/VariantService.java
@@ -1,8 +1,6 @@
 package com.example.fastest_server.variant;
 
-import com.example.fastest_server.question.DocxReader;
 import lombok.AllArgsConstructor;
-import org.docx4j.openpackaging.exceptions.Docx4JException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,10 +10,10 @@ import java.util.List;
 public class VariantService {
     private final VariantRepository variantRepository;
 
-    public void setMarkToVariant (int idVariant,int mark){
+    public void setMarkToVariant (Long idVariant, Integer mark){
         variantRepository.setMarkToVariant(idVariant,mark);
     };
-    public List<Variant> getVariantsList(int idTest)  {
+    public List<Variant> getVariantsList(Long idTest)  {
         return variantRepository.findByTestId(idTest);
     }
 }
